After all, add inputs_seq_len

+++ utils/dataset/all_load/attention_all_load.py
@@ -29,11 +29,16 @@ class DatasetBase(Base):
             batch_size (int, optional): the size of mini-batch
         Returns:
             A tuple of `(inputs, labels, inputs_seq_len, labels_seq_len, input_names)`
-                inputs: list of input data of size `[B, T, input_dim]`
-                labels: list of target labels of size `[B, T]`
-                inputs_seq_len: list of length of inputs of size `[B]`
-                labels_seq_len: list of length of target labels of size `[B]`
-                input_names: list of file name of input data of size `[B]`
+                inputs: list of input data of size
+                    `[B, T, input_dim]`
+                labels: list of target labels of size
+                    `[B, T]`
+                inputs_seq_len: list of length of inputs of size
+                    `[B]`
+                labels_seq_len: list of length of target labels of size
+                    `[B]`
+                input_names: list of file name of input data of size
+                    `[B]`
             is_new_epoch: If true, one epoch is finished
         """
         if self.max_epoch is not None and self.epoch >= self.max_epoch:
@@ -102,13 +107,12 @@ class DatasetBase(Base):
         # Initialization
         inputs = np.zeros(
             (len(data_indices), max_frame_num,
-             self.input_list[0].shape[-1] * self.splice),
-            dtype=np.float32)
+             self.input_list[0].shape[-1] * self.splice), dtype=np.float32)
         labels = np.array(
             [[self.padded_value] * max_seq_len] * len(data_indices),
             dtype=np.int32)
         inputs_seq_len = np.zeros((len(data_indices),), dtype=np.int32)
-        labels_seq_len = np.zeros((len(data_indices),), dtype=int)
+        labels_seq_len = np.zeros((len(data_indices),), dtype=np.int32)
         input_names = np.array(list(
             map(lambda path: basename(path).split('.')[0],
                 np.take(self.input_paths, data_indices, axis=0))))
@@ -130,5 +134,6 @@ class DatasetBase(Base):
             labels_seq_len[i_batch] = len(self.label_list[x])
 
         self.iteration += len(data_indices)
+
         return (inputs, labels, inputs_seq_len, labels_seq_len,
                 input_names), self.is_new_epoch
